Proxy-rec
Hello
The script I wrote uses an API to receive proxies (Http, Socks4, Socks5) and after receiving it, it starts checking the proxies against the address www.google.com and after checking the proxies in two files. Saves Proxy.txt and ProxyCheck.txt. The proxies received directly from the API are saved in the Proxy.txt file, and the checked proxies are saved in the ProxyCheck.txt file.